refactor: rename dir MCQ to MCQSection
