+++ src/components/Academics/MCQSection/index.tsx
